[SYCL][NFC] Do not use std::exclusive_scan/inclusive_scan/reduce in tests
