+++ sycl/test-e2e/GroupAlgorithm/exclusive_scan_sycl2020.cpp
@@ -2,6 +2,7 @@
 // RUN: %{build} -fsycl-device-code-split=per_kernel -I . -o %t.out
 // RUN: %{run} %t.out
 
+#include "../helpers.hpp"
 #include "support.h"
 #include <algorithm>
 #include <cassert>
@@ -15,24 +16,6 @@ using namespace sycl;
 
 template <class SpecializationKernelName, int TestNumber>
 class exclusive_scan_kernel;
-
-// std::exclusive_scan isn't implemented yet, so use serial implementation
-// instead
-// TODO: use std::exclusive_scan when it will be supported
-namespace emu {
-template <typename InputIterator, typename OutputIterator,
-          class BinaryOperation, typename T>
-OutputIterator exclusive_scan(InputIterator first, InputIterator last,
-                              OutputIterator result, T init,
-                              BinaryOperation binary_op) {
-  T partial = init;
-  for (InputIterator it = first; it != last; ++it) {
-    *(result++) = partial;
-    partial = binary_op(partial, *it);
-  }
-  return result;
-}
-} // namespace emu
 
 queue q;
