+++ sycl/test-e2e/Regression/exclusive-scan-char-short.cpp
@@ -4,6 +4,7 @@
 // This test ensures the result computed by exclusive_scan_over_group
 // for the first work item when given a short or char argument with
 // the maximum or minimum operator is computed correctly.
+#include "../helpers.hpp"
 #include <numeric>
 #include <sycl/sycl.hpp>
 
@@ -18,7 +19,7 @@ template <typename T, typename OpT> void test() {
   auto *p = malloc_shared<T>(1, q);
   *p = 0;
   T ref;
-  std::exclusive_scan(p, p + 1, &ref, init, op);
+  emu::exclusive_scan(p, p + 1, &ref, init, op);
   range r(1);
   q.parallel_for(nd_range(r, r), [=](nd_item<1> it) {
      auto g = it.get_group();
